Switch file endings on test files, add utils to test directory.
