+++ tests/testfiles/both-arg-types.js
@@ -1,0 +1,7 @@
+const runnableExports = require('../../src/index.js');
+
+module.exports.test = (objectArg, ...arrayArgs) => {
+	console.log(objectArg, arrayArgs);
+};
+
+runnableExports();
