--- tests/testfiles/both-arg-types.spec.js
@@ -1,7 +1,0 @@
-const runnableExports = require('../../src/index.js')
-
-module.exports.test = (objectArg, ...arrayArgs) => {
-	console.log(objectArg, arrayArgs)
-}
-
-runnableExports()
