+++ tests/testfiles/defaultexport.js
@@ -1,0 +1,7 @@
+const runnableExports = require('../../src/index.js');
+
+module.exports = (...testArgs) => {
+	console.log(testArgs);
+};
+
+runnableExports();
