--- tests/testfiles/defaultexport.spec.js
@@ -1,7 +1,0 @@
-const runnableExports = require('../../src/index.js')
-
-module.exports = (...testArgs) => {
-	console.log(testArgs)
-}
-
-runnableExports()
