--- tests/testfiles/export.spec.js
@@ -1,7 +1,0 @@
-const runnableExports = require('../../src/index.js')
-
-module.exports.test = (...testArgs) => {
-	console.log(testArgs)
-}
-
-runnableExports();
