+++ tests/testfiles/nested.js
@@ -1,0 +1,3 @@
+require('./export');
+
+console.log('cache deleted');
