--- tests/testfiles/nested.spec.js
@@ -1,2 +1,0 @@
-require('./export.spec')
-console.log('cache deleted')
