+++ tests/testfiles/noexport.js
@@ -1,0 +1,5 @@
+const runnableExports = require('../../src/index.js');
+
+module.exports = {};
+
+runnableExports();
