--- tests/testfiles/noexport.spec.js
@@ -1,5 +1,0 @@
-const runnableExports = require('../../src/index.js')
-
-module.exports = {}
-
-runnableExports()
